Not to install twice the font

+++ setup/mac/font.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 
-cp -f $(brew --prefix)/share/fonts/Ricty*.ttf ~/Library/Fonts/
-fc-cache -vf
-echo "Ricty font was installed"
+if [ ! -f ~/Library/Fonts/Ricty-Regular.ttf ]; then
+  cp -f $(brew --prefix)/share/fonts/Ricty*.ttf ~/Library/Fonts/
+  fc-cache -vf
+  echo "Ricty font was installed"
+fi
